requirements.txt
# API / Web server
fastapi
uvicorn
websockets

# Audio handling
faster-whisper
sounddevice

# Testing
pytest

# LLM
google-genai
pydantic
openai
azure-ai-contentsafety
azure-identity
